+++ Changelog.md
@@ -1,7 +1,8 @@
 Changelog
 =========
 
+
 0.1.0 (2016-07-06)
---------------------
+------------------
 * [Feature] Implemented basic functionality: init, messages, release, append, current, edit